Revert "basic additions for 2 distributed kuka robots (#1)"
This reverts commit d32f6a66ccacfeaa02b4eba8e61734da6c6269f9.
