+++ kuka_kr16/kuka_kr16_bringup/kuka_kr16_2_bringup/controller_config/kuka_6dof_forward_position_controller_goals.yaml
@@ -1,0 +1,10 @@
+publisher_forward_position_controller:
+  ros__parameters:
+    wait_sec_between_publish: 5
+    publish_topic: /forward_position_controller/commands
+
+    goal_names: ["pos1", "pos2", "pos3", "pos4"]
+    pos1: [0.785, 0.785, 0.6, 0.785, 0.785, 0.785]
+    pos2: [0, 0, 0, 0, 0, 0]
+    pos3: [-0.785, -0.785, -0.785, -0.785, -0.785, -0.785]
+    pos4: [0, 0, 0, 0, 0, 0]
