+++ kuka_kr16/kuka_kr16_bringup/kuka_kr16_2_bringup/launch/kuka_kr16_2.launch.py
@@ -1,0 +1,141 @@
+import os
+
+from ament_index_python.packages import get_package_share_directory
+
+from launch import LaunchDescription
+from launch.actions import DeclareLaunchArgument, RegisterEventHandler
+from launch.event_handlers import OnProcessExit
+from launch.substitutions import Command, FindExecutable, LaunchConfiguration, PathJoinSubstitution
+
+from launch_ros.actions import Node
+from launch_ros.substitutions import FindPackageShare
+
+import xacro
+
+
+def generate_launch_description():
+
+    declared_arguments = []
+
+    declared_arguments.append(
+        DeclareLaunchArgument(
+            "prefix",
+            default_value="",
+            description="Prefix of the joint names, useful for \
+        multi-robot setup. If changed than also joint names in the controllers' configuration \
+        have to be updated.",
+        )
+    )
+    declared_arguments.append(
+        DeclareLaunchArgument(
+            "use_mock_hw",
+            default_value="true",
+            description="Start robot with fake hardware mirroring command to its states."
+        )
+    )
+
+    declared_arguments.append(
+        DeclareLaunchArgument(
+            "use_mock_sensor_commands",
+            default_value="false",
+            description="Enable fake command interfaces for sensors used for simple simulations. \
+            Used only if 'use_mock_hardware' parameter is true.",
+        )
+    )
+
+    # initialize arguments
+    prefix = LaunchConfiguration("prefix")
+    use_mock_hw = LaunchConfiguration("use_mock_hw")
+    use_mock_sensor_commands = LaunchConfiguration("use_mock_sensor_commands")
+
+    robot_description_content = Command(
+        [
+            PathJoinSubstitution([FindExecutable(name="xacro")]),
+            " ",
+            PathJoinSubstitution(
+                [FindPackageShare("kuka_kr16_2_description"),
+                 "urdf", "kuka_kr16.urdf.xacro"]
+            ),
+            " ",
+            "prefix:=",
+            prefix,
+            " ",
+            "use_mock_hw:=",
+            use_mock_hw,
+            " ",
+            "use_mock_sensor_commands:=",
+            use_mock_sensor_commands,
+            " ",
+        ]
+    )
+
+    robot_description = {"robot_description": robot_description_content}
+
+    robot_controllers = PathJoinSubstitution(
+        [FindPackageShare("kuka_kr16_2_bringup"), "controller_config",
+         "kuka_6dof_controllers.yaml"]
+    )
+
+    control_node = Node(
+        package="controller_manager",
+        executable="ros2_control_node",
+        output="both",
+        parameters=[robot_description, robot_controllers],
+    )
+
+    robot_state_pub_node = Node(
+        package="robot_state_publisher",
+        executable="robot_state_publisher",
+        parameters=[robot_description],
+        output="both",
+    )
+
+    joint_state_broadcaster_spawner = Node(
+        package="controller_manager",
+        executable="spawner",
+        arguments=["joint_state_broadcaster",
+                   "--controller-manager", "/controller_manager"],
+    )
+
+    robot_controller_spawner = Node(
+        package="controller_manager",
+        executable="spawner",
+        arguments=["forward_position_controller", "-c", "/controller_manager"],
+    )
+
+    delay_robot_controller_spawner_after_joint_state_broadcaster_spawner = RegisterEventHandler(
+        event_handler=OnProcessExit(
+            target_action=joint_state_broadcaster_spawner,
+            on_exit=[robot_controller_spawner],
+        )
+    )
+
+    rviz_config_file = PathJoinSubstitution(
+        [FindPackageShare("kuka_kr16_2_description"),
+         "rviz_config", "kuka_kr16_2.rviz"]
+    )
+
+    rviz_node = Node(
+        package="rviz2",
+        executable="rviz2",
+        name="rviz2",
+        output="log",
+        arguments=["-d", rviz_config_file],
+    )
+
+    delay_rviz_after_joint_state_broadcaster_spawner = RegisterEventHandler(
+        event_handler=OnProcessExit(
+            target_action=joint_state_broadcaster_spawner,
+            on_exit=[rviz_node],
+        )
+    )
+
+    nodes = [
+        control_node,
+        robot_state_pub_node,
+        joint_state_broadcaster_spawner,
+        delay_rviz_after_joint_state_broadcaster_spawner,
+        delay_robot_controller_spawner_after_joint_state_broadcaster_spawner
+    ]
+
+    return LaunchDescription(declared_arguments + nodes)
