+++ kuka_kr16/kuka_kr16_description/kuka_kr16_2_description/package.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0"?>
+<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
+<package format="3">
+  <name>kuka_kr16_2_description</name>
+  <version>0.0.0</version>
+  <description>Package with URDF and description files for the kuka kr 16_2 robot.</description>
+
+  <maintainer email="manuel.muth@stoglrobotics.de">Manuel Muth</maintainer>
+
+  <license>Apache-2.0</license>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+
+  <exec_depend>joint_state_publisher_gui</exec_depend>
+  <exec_depend>robot_state_publisher</exec_depend>
+  <exec_depend>rviz2</exec_depend>
+
+  <export>
+    <build_type>ament_cmake</build_type>
+  </export>
+</package>
