--- two_distributed_kukas/two_distributed_kukas_bringup/CMakeLists.txt
@@ -1,21 +1,0 @@
-cmake_minimum_required(VERSION 3.5)
-project(two_distributed_kukas_bringup)
-
-# Default to C++14
-if(NOT CMAKE_CXX_STANDARD)
-  set(CMAKE_CXX_STANDARD 14)
-endif()
-
-if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
-  add_compile_options(-Wall -Wextra -Wpedantic)
-endif()
-
-# find dependencies
-find_package(ament_cmake REQUIRED)
-
-install(
-  DIRECTORY controller_config launch rviz_config
-  DESTINATION share/${PROJECT_NAME}
-)
-
-ament_package()
