--- two_distributed_kukas/two_distributed_kukas_bringup/controller_config/first_distributed_kuka_kr16_2_controller.yaml
@@ -1,9 +1,0 @@
-/sub_1/controller_manager:
-  ros__parameters:
-    update_rate: 10  # Hz
-    distributed: true
-    sub_controller_manager: true
-    distributed_interfaces_publish_period: 12
-
-    joint_state_broadcaster:
-      type: joint_state_broadcaster/JointStateBroadcaster
