--- two_distributed_kukas/two_distributed_kukas_bringup/controller_config/joint_trajectory_goals/one_central_one_distributed_kuka_6dof_joint_trajectory_controller_goals.yaml
@@ -1,39 +1,0 @@
-publisher_joint_trajectory_controller:
-  ros__parameters:
-    controller_name: "position_trajectory_controller"
-    wait_sec_between_publish: 6
-
-    goal_names: ["pos1", "pos2", "pos3", "pos4"]
-    pos1: [0.0, -1.57, 1.57, 0.0, 1.57, 0.0, 0.0, -1.57, 1.57, 0.0, 1.57, 0.0]
-    pos2: [0.0, -1.57, 1.57, 0.0, 0.0, 0.0, 0.0, -1.57, 1.57, 0.0, 0.0, 0.0]
-    pos3: [0.0, -1.57, 1.57, 0.0, 1.57, 0.0, 0.0, -1.57, 1.57, 0.0, 1.57, 0.0]
-    pos4: [0.0, -1.57, 1.57, 0.0, 0.0, 0.0, 0.0, -1.57, 1.57, 0.0, 0.0, 0.0]
-
-    joints:
-      - joint_a1
-      - joint_a2
-      - joint_a3
-      - joint_a4
-      - joint_a5
-      - joint_a6
-      - /sub_1/sub_1_joint_a1
-      - /sub_1/sub_1_joint_a2
-      - /sub_1/sub_1_joint_a3
-      - /sub_1/sub_1_joint_a4
-      - /sub_1/sub_1_joint_a5
-      - /sub_1/sub_1_joint_a6
-
-    # check_starting_point: false
-    # starting_point_limits:
-    #   joint_a1: [-0.1,0.1]
-    #   joint_a2: [-0.1,0.1]
-    #   joint_a3: [-0.1,0.1]
-    #   joint_a4: [-0.1,0.1]
-    #   joint_a5: [-0.1,0.1]
-    #   joint_a6: [-0.1,0.1]
-    #   /sub_1/sub_1_joint_a1: [-0.1,0.1]
-    #   /sub_1/sub_1_joint_a2: [-0.1,0.1]
-    #   /sub_1/sub_1_joint_a3: [-0.1,0.1]
-    #   /sub_1/sub_1_joint_a4: [-0.1,0.1]
-    #   /sub_1/sub_1_joint_a5: [-0.1,0.1]
-    #   /sub_1/sub_1_joint_a6: [-0.1,0.1]
