--- two_distributed_kukas/two_distributed_kukas_bringup/controller_config/one_central_one_distributed_aio.yaml
@@ -1,45 +1,0 @@
-controller_manager:
-  ros__parameters:
-    distributed: true
-    sub_controller_manager: false # false -> is centralized controller.
-
-    joint_state_broadcaster:
-      type: joint_state_broadcaster/JointStateBroadcaster
-
-    forward_position_controller:
-      type: forward_command_controller/ForwardCommandController
-
-    position_trajectory_controller:
-      type: joint_trajectory_controller/JointTrajectoryController
-
-position_trajectory_controller:
-  ros__parameters:
-    joints:
-      - joint_a1
-      - joint_a2
-      - joint_a3
-      - joint_a4
-      - joint_a5
-      - joint_a6
-      - /sub_1/sub_1_joint_a1
-      - /sub_1/sub_1_joint_a2
-      - /sub_1/sub_1_joint_a3
-      - /sub_1/sub_1_joint_a4
-      - /sub_1/sub_1_joint_a5
-      - /sub_1/sub_1_joint_a6
-
-    command_interfaces:
-      - position
-
-    state_interfaces:
-      - position
-
-    state_publish_rate: 2.0 # Defaults to 50
-    action_monitor_rate: 2.0 # Defaults to 20
-
-    allow_partial_joints_goal: false # Defaults to false
-    open_loop_control: true
-    allow_integration_in_goal_trajectories: true
-    constraints:
-      stopped_velocity_tolerance: 0.01 # Defaults to 0.01
-      goal_time: 0.0 # Defaults to 0.0 (start immediately)
