--- two_distributed_kukas/two_distributed_kukas_bringup/package.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>two_distributed_kukas_bringup</name>
-  <version>0.0.0</version>
-  <description>Package with launch files and run-time configurations for  my master thesis examples using simply rrbots.</description>
-
-  <maintainer email="manuel.muth@stoglrobotics.de">Manuel Muth</maintainer>
-
-  <license>Apache-2.0</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <exec_depend>forward_command_controller</exec_depend>
-  <exec_depend>joint_state_broadcaster</exec_depend>
-  <exec_depend>joint_state_publisher_gui</exec_depend>
-  <exec_depend>joint_trajectory_controller</exec_depend>
-  <exec_depend>robot_state_publisher</exec_depend>
-  <exec_depend>rviz2</exec_depend>
-  <exec_depend>two_distributed_rrbots_hw</exec_depend>
-  <exec_depend>xacro</exec_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
