--- two_distributed_rrbots/two_distributed_rrbots_bringup/controller_config/kuka_6dof_joint_trajectory_controller_goals.yaml
@@ -1,18 +1,0 @@
-publisher_joint_trajectory_controller:
-  ros__parameters:
-    controller_name: "position_trajectory_controller"
-    wait_sec_between_publish: 5
-
-    goal_names: ["pos1", "pos2", "pos3", "pos4"]
-    pos1: [0.785, 0.785, 0.2, 0.2, 0.785, 0.785]
-    pos2: [0, 0, 0, 0, 0, 0]
-    pos3: [-0.785, -0.785, -0.785, -0.785, -0.785, -0.785]
-    pos4: [0, 0, 0, 0, 0, 0]
-    
-    joints:
-      - joint1
-      - joint2
-      - /sub_1/sub_1_joint1
-      - /sub_1/sub_1_joint2
-      - /sub_2/sub_2_joint1
-      - /sub_2/sub_2_joint2
