+++ two_distributed_rrbots/two_distributed_rrbots_bringup/controller_config/rrbot_controllers_satellite_1.yaml
@@ -3,8 +3,6 @@
     update_rate: 10  # Hz
     distributed: true
     sub_controller_manager: true
-    use_multiple_nodes: true
-    distributed_interfaces_publish_period: 12
 
     joint_state_broadcaster:
       type: joint_state_broadcaster/JointStateBroadcaster
