+++ two_distributed_rrbots/two_distributed_rrbots_bringup/controller_config/rrbot_controllers_satellite_2.yaml
@@ -3,7 +3,6 @@
     update_rate: 10  # Hz
     distributed: true
     sub_controller_manager: true
-    distributed_interfaces_publish_period: 12
 
     joint_state_broadcaster:
       type: joint_state_broadcaster/JointStateBroadcaster
