--- two_distributed_rrbots/two_distributed_rrbots_bringup/launch/joint_trajectory_controller_goal_publisher.launch.py
@@ -1,41 +1,0 @@
-# Copyright 2022 Stogl Robotics Consulting UG (haftungsbeschränkt)
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from launch import LaunchDescription
-from launch.substitutions import PathJoinSubstitution
-from launch_ros.actions import Node
-from launch_ros.substitutions import FindPackageShare
-
-
-def generate_launch_description():
-
-    position_goals = PathJoinSubstitution(
-        [
-            FindPackageShare("two_distributed_rrbots_bringup"),
-            "controller_config",
-            "kuka_6dof_joint_trajectory_controller_goals.yaml",
-        ]
-    )
-
-    return LaunchDescription(
-        [
-            Node(
-                package="ros2_controllers_test_nodes",
-                executable="publisher_joint_trajectory_controller",
-                name="publisher_joint_trajectory_controller",
-                parameters=[position_goals],
-                output="both",
-            )
-        ]
-    )
