+++ two_distributed_rrbots/two_distributed_rrbots_bringup/launch/two_distributed_rrbots.launch.py
@@ -23,7 +23,7 @@ from launch_ros.substitutions import FindPackageShare
 
 
 def delete_direct_slash_duplicate(t):
-    if not (t[0] == "/" and t[1] == "/"):
+    if (not (t[0] == "/" and t[1] == "/")):
         return t[0]
 
 
@@ -32,12 +32,7 @@ def prepend_slash_if_not_null(prefix):
         return ""
     ns = "/" + prefix
     # remove all occurrences of slashes that directly follow each other ("//Prefix/////Namespace//" -> "/Prefix/Namespace/")
-    return "".join(
-        filter(
-            lambda item: item is not None,
-            map(delete_direct_slash_duplicate, zip(ns, ns[1:] + " ")),
-        )
-    )
+    return ''.join(filter(lambda item: item is not None, map(delete_direct_slash_duplicate, zip(ns, ns[1:] + " "))))
 
 
 def generate_launch_description():
@@ -46,15 +41,14 @@ def generate_launch_description():
     controller_manager_name = "controller_manager"
     satellite_1_ns_name = "sub_1"
     satellite_2_ns_name = "sub_2"
-    slash_controller_manager_name = prepend_slash_if_not_null(controller_manager_name)
+    slash_controller_manager_name = prepend_slash_if_not_null(
+        controller_manager_name)
     slash_satellite_1_ns_name = prepend_slash_if_not_null(satellite_1_ns_name)
     slash_satellite_2_ns_name = prepend_slash_if_not_null(satellite_2_ns_name)
-    satellite_1_controller_manager_name = (
-        slash_satellite_1_ns_name + slash_controller_manager_name
-    )
-    satellite_2_controller_manager_name = (
-        slash_satellite_2_ns_name + slash_controller_manager_name
-    )
+    satellite_1_controller_manager_name = slash_satellite_1_ns_name + \
+        slash_controller_manager_name
+    satellite_2_controller_manager_name = slash_satellite_2_ns_name + \
+        slash_controller_manager_name
 
     # Get URDF via xacro
     main_robot_description_content = Command(
@@ -85,8 +79,7 @@ def generate_launch_description():
                 ]
             ),
             " ",
-            "prefix:=",
-            satellite_1_ns_name + "_",
+            "prefix:=sub_1_",
             " ",
             "base_color:=green",
         ]
@@ -104,21 +97,19 @@ def generate_launch_description():
                 ]
             ),
             " ",
-            "prefix:=",
-            satellite_2_ns_name + "_",
+            "prefix:=sub_2_",
             " ",
             "base_color:=blue",
         ]
     )
 
     # create parameter from description content
-    main_robot_description = {"robot_description": main_robot_description_content}
+    main_robot_description = {
+        "robot_description": main_robot_description_content}
     robot_satellite_1_description = {
-        "robot_description": robot_satellite_1_description_content
-    }
+        "robot_description": robot_satellite_1_description_content}
     robot_satellite_2_description = {
-        "robot_description": robot_satellite_2_description_content
-    }
+        "robot_description": robot_satellite_2_description_content}
 
     # Get controller manager settings
     main_robot_controllers = PathJoinSubstitution(
@@ -176,15 +167,13 @@ def generate_launch_description():
     robot_controller_spawner = Node(
         package="controller_manager",
         executable="spawner",
-        arguments=["position_trajectory_controller", "-c", "/controller_manager"],
+        arguments=["forward_position_controller", "-c", "/controller_manager"],
     )
 
-    delay_robot_controller_spawner_after_joint_state_broadcaster_spawner = (
-        RegisterEventHandler(
-            event_handler=OnProcessExit(
-                target_action=joint_state_broadcaster_spawner,
-                on_exit=[robot_controller_spawner],
-            )
+    delay_robot_controller_spawner_after_joint_state_broadcaster_spawner = RegisterEventHandler(
+        event_handler=OnProcessExit(
+            target_action=joint_state_broadcaster_spawner,
+            on_exit=[robot_controller_spawner],
         )
     )
 
@@ -194,16 +183,20 @@ def generate_launch_description():
         package="controller_manager",
         executable="ros2_control_node",
         namespace=satellite_1_ns_name,
-        parameters=[robot_satellite_1_description, robot_controllers_satellite_1],
+        parameters=[robot_satellite_1_description,
+                    robot_controllers_satellite_1],
         remappings=[
             (
                 "/forward_position_controller/commands",
                 "/position_commands",
             ),
-            ("/joint_states", slash_satellite_1_ns_name + "/joint_states"),
+            (
+                "/joint_states",
+                slash_satellite_1_ns_name + "/joint_states"
+            ),
             (
                 "/dynamic_joint_states",
-                slash_satellite_1_ns_name + "/dynamic_joint_states",
+                slash_satellite_1_ns_name + "/dynamic_joint_states"
             ),
         ],
         output="both",
@@ -221,11 +214,24 @@ def generate_launch_description():
         package="controller_manager",
         executable="spawner",
         namespace=satellite_1_ns_name,
-        arguments=[
-            "joint_state_broadcaster",
-            "-c",
-            satellite_1_controller_manager_name,
-        ],
+        arguments=["joint_state_broadcaster", "-c",
+                   satellite_1_controller_manager_name],
+    )
+
+    robot_controller_spawner_1 = Node(
+        package="controller_manager",
+        executable="spawner",
+        namespace=satellite_1_ns_name,
+        arguments=["forward_position_controller",
+                   "-c", satellite_1_controller_manager_name],
+    )
+
+    # Delay start of robot_controller after `joint_state_broadcaster`
+    delay_robot_controller_spawner_after_joint_state_broadcaster_spawner_1 = RegisterEventHandler(
+        event_handler=OnProcessExit(
+            target_action=joint_state_broadcaster_spawner_1,
+            on_exit=[robot_controller_spawner_1],
+        )
     )
 
     # SUBSYSTEMS
@@ -235,16 +241,20 @@ def generate_launch_description():
         package="controller_manager",
         executable="ros2_control_node",
         namespace=satellite_2_ns_name,
-        parameters=[robot_satellite_2_description, robot_controllers_satellite_2],
+        parameters=[robot_satellite_2_description,
+                    robot_controllers_satellite_2],
         remappings=[
             (
                 "/forward_position_controller/commands",
                 "/position_commands",
             ),
-            ("/joint_states", slash_satellite_2_ns_name + "/joint_states"),
+            (
+                "/joint_states",
+                slash_satellite_2_ns_name + "/joint_states"
+            ),
             (
                 "/dynamic_joint_states",
-                slash_satellite_2_ns_name + "/dynamic_joint_states",
+                slash_satellite_2_ns_name + "/dynamic_joint_states"
             ),
         ],
         output="both",
@@ -262,20 +272,30 @@ def generate_launch_description():
         package="controller_manager",
         executable="spawner",
         namespace=satellite_2_ns_name,
-        arguments=[
-            "joint_state_broadcaster",
-            "-c",
-            satellite_2_controller_manager_name,
-        ],
+        arguments=["joint_state_broadcaster", "-c",
+                   satellite_2_controller_manager_name],
+    )
+
+    robot_controller_spawner_2 = Node(
+        package="controller_manager",
+        executable="spawner",
+        namespace=satellite_2_ns_name,
+        arguments=["forward_position_controller",
+                   "-c", satellite_2_controller_manager_name],
+    )
+
+    # Delay start of robot_controller after `joint_state_broadcaster`
+    delay_robot_controller_spawner_after_joint_state_broadcaster_spawner_2 = RegisterEventHandler(
+        event_handler=OnProcessExit(
+            target_action=joint_state_broadcaster_spawner_2,
+            on_exit=[robot_controller_spawner_2],
+        )
     )
 
     # RVIZ
     rviz_config_file = PathJoinSubstitution(
-        [
-            FindPackageShare("two_distributed_rrbots_description"),
-            "rviz_config",
-            "two_distributed_rrbots_all_in_on.rviz",
-        ]
+        [FindPackageShare("two_distributed_rrbots_description"),
+         "rviz_config", "two_distributed_rrbots_all_in_on.rviz"]
     )
 
     rviz_node = Node(
@@ -302,9 +322,11 @@ def generate_launch_description():
         sub_1_control_node,
         robot_state_pub_node_1,
         joint_state_broadcaster_spawner_1,
+        # delay_robot_controller_spawner_after_joint_state_broadcaster_spawner_1,
         sub_2_control_node,
         robot_state_pub_node_2,
         joint_state_broadcaster_spawner_2,
+        # delay_robot_controller_spawner_after_joint_state_broadcaster_spawner_2,
         delay_rviz_after_joint_state_broadcaster_spawner,
     ]
